working on MaskedNMA
